Update README.md
📝 About the Project

Spam emails are annoying and sometimes dangerous. In this project, I built a machine learning model that can automatically classify messages as **spam** or **not spam (ham)** using natural language processing (NLP) techniques.

The goal is to demonstrate how basic text classification works using real-world data. This project uses the **Naive Bayes algorithm** along with **CountVectorizer** to convert text into numerical features and make predictions.

This project helped me strengthen my understanding of:

- Text data preprocessing  
- Feature extraction from natural language  
- Applying classification algorithms to real-world datasets  
- Evaluating model accuracy and performance

It is developed as part of the **DataZenix Internship Program** to gain hands-on experience in Data Science and Machine Learning.